possession update / fix
Moved possession onto the player controller (out of the player character) and removed possession code from the PropPawn. However, added a variable and events to PropPawn to keep track of if it is currently possessed. This is used by the player controller to check if we can possess it or not.